+++ packages/gfc/src/utils/p.ts
@@ -8,7 +8,11 @@
  * @param key - a property key
  * @param value - a property value
  */
-export function setProp<P, Key extends keyof P = keyof P>(this: { props: P; isConnected: boolean }, key: Key, value: P[Key]) {
+export function setProp<P, Key extends keyof P = keyof P>(
+  this: { props: P; isConnected: boolean },
+  key: Key,
+  value: P[Key],
+) {
   if (!this.isConnected) {
     this.props[key] = value;
     return;
@@ -16,8 +20,8 @@ export function setProp<P, Key extends keyof P = keyof P>(this: { props: P; isCo
 
   if (value !== this.props[key]) {
     // Stash changed properties between updates
-    this['__stash__'] = {
-      ...this['__stash__'],
+    this["__stash__"] = {
+      ...this["__stash__"],
       [key]: value,
     };
   }
@@ -25,7 +29,7 @@ export function setProp<P, Key extends keyof P = keyof P>(this: { props: P; isCo
   Promise.resolve({
     then: () => {
       // Stashed changed properties
-      let stash: Partial<P> | null = this['__stash__'];
+      let stash: Partial<P> | null = this["__stash__"];
 
       if (stash != null) {
         this.props = {
@@ -44,14 +48,15 @@ export const p = {
     return {
       type: {} as Type,
       attribute,
-    };
+      isReq: true,
+    } as const;
   },
 
   opt<Type>(attribute?: string) {
     return {
       type: {} as Type | undefined,
       attribute,
-      optional: true,
-    };
+      isReq: false,
+    } as const;
   },
 };